requirements.txt
fastapi
uvicorn
python-multipart
python-docx
pandas
openpyxl
xlsxwriter
requests
pydantic
python-dotenv
textract
pdfplumber
